Fix pointer arithmetic in regular expression handling
`*str++` was returned in several places while the `trex_matchnode()`
function is expected to return a pointer to `TRexChar` (and `str` is
also a pointer to `TRexChar`). So, we fix that by returning `str++`
instead.

Moreover, a pair of curly braces make the `OP_DOT` matcher not return
the appropriate value.