build: remove _static postfix from static lib
Originally, the _static postfix is designed for Win32, so that we can
differentiate between the static library and the import library.
However, such design will make linking in UNIX environments, such as
Linux, unnecessarily more complicated.

This patch removes the _static postfix from the static library, so in
UNIX environments, we can just use -largtable3.

Resolves #101