fix: add context to command handler callback
The current design does not allow passing user-defined data to the
command handler callback, which makes it difficult to shared data
between the main program and callbacks.

In this patch, we introduce an API change, which may break the code that
use the sub-command mechanism. However, since the sub-command API has
not been publicly introduced, we hope this necessary change will not
have big impact to users.