build: refactor makefile and remove bundled win32 executables
This patch removes the Win32 executables from the tools directory.
Instead of bundling build tools, we now use a Docker container to
run the build script in an Ubuntu environment, even on Windows.

Additionally, the Makefile has been refactored to allow generating
release archives with a single command, simplifying the build process.