fix: update version.tag to follow semver and revise cmake version parsing
This commit updates `version.tag` to adhere to the SemVer standard and
modifies the main `CMakeLists.txt` to parse the updated version string
correctly.

Additionally, `PROJECT_VERSION` is redefined to align with CMake's
default definition, which includes `PROJECT_VERSION_TWEAK`. As a result,
other version constants are updated to reflect this change.